Implements `loadMessage` func with Operation
This function should fetch both parts of the message (concurrently using GCD or OperationQueue) and join them with
a space. e.g if `fetchMessageOne` completes with "Good" and `fetchMessageTwo` completes with "morning!" then loadMessage should call it's completion once with the String:
  "Good morning!"
If loading either part of the message takes more than 2 seconds then `loadMessage` should complete with the String
  "Unable to load message - Time out exceeded"
The function should only complete once and must always return the full message in the correct order.